+++ mastodon-s3/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Changes to fix broken dependancies
 * Yarn flag change from `yarn install --pure-lockfile` to `yarn install --immutable`
+* Manually add yarn packages `babel-plugin-lodash@3.3.4` and `compression-webpack-plugin@10.0.0` to fix a compile error
 
 ---